@@ -1,7 +1,7 @@
 window.addEventListener("load", function() {
     (function ($) {
         "use strict";
-    
+
         function initItem(item) {
             var empty_label, chainfield = "#id_" + $(item).attr("data-chainfield"),
                 url = $(item).attr("data-url"),
@@ -20,7 +20,7 @@ window.addEventListener("load", function() {
                 chainedm2m.init(chainfield, url, id, value, auto_choose);
             }
         }
-    
+
         $(window).on('load', function () {
             $.each($(".chained"), function (index, item) {
                 initItem(item);
@@ -31,13 +31,13 @@ window.addEventListener("load", function() {
                 }
             });
         });
-    
+
         $(document).ready(function () {
             $.each($(".chained-fk"), function (index, item) {
                 initItem(item);
             });
         });
-    
+
         function initFormset(chained) {
             var re = /\d+/g,
                 prefix,
@@ -55,34 +55,34 @@ window.addEventListener("load", function() {
                         prefix = match[0];
                     }
                 } while (match);
-    
+
                 chainfield = chainfield.replace("__prefix__", prefix);
                 $(chained).attr("data-chainfield", chainfield);
             }
             initItem(chained);
         }
-    
-        $(document).on('formset:added', function (event) {
-            let row = $('#'+event.target.id)
+
+        $(document).on('formset:added', function (event, _row) {
             // Fired every time a new inline formset is created
-    
+
             var chainedFK, chainedM2M, filteredM2M;
-    
+            var $row = _row || $(event.target);
+
             // For the ForeingKey
-            chainedFK = row.find(".chained-fk");
+            chainedFK = $row.find(".chained-fk");
             $.each(chainedFK, function (index, chained) {
                 initFormset(chained);
             });
-    
+
             // For the ManyToMany
-            chainedM2M = row.find(".chained");
+            chainedM2M = $row.find(".chained");
             $.each(chainedM2M, function (index, chained) {
                 initFormset(chained);
             });
-    
+
             // For the ManyToMany using horizontal=True added after the page load
             // using javascript.
-            filteredM2M = row.find(".filtered");
+            filteredM2M = $row.find(".filtered");
             $.each(filteredM2M, function (index, filtered) {
                 if (filtered.hasAttribute('data-chainfield')) {
                     initFormset(filtered);
